Feature: configured pytest and created example test

+++ myrestaurant/myrestaurant_app/models.py
@@ -6,7 +6,7 @@ class Inventory(models.Model):
     id = models.BigAutoField(primary_key=True)
     ingredient = models.CharField(max_length=30)
     slug = models.SlugField(unique=True, max_length=30, blank=True)
-    quantity = models.IntegerField(default=0)
+    quantity = models.PositiveSmallIntegerField(default=0)
     unit_price = models.DecimalField(max_digits=5, decimal_places=2, default=None, blank=True)
     image = models.ImageField(upload_to='inventory', blank=True)
